+++ src/filterfir.cpp
@@ -27,13 +27,13 @@
 
 namespace dsignal {
 
-FilterFir::FilterFir(int max_buffer_size, bool throw_overflow):
+FilterFir::FilterFir(size_t max_buffer_size, bool throw_overflow):
     SignalProcessorBuffered(max_buffer_size, throw_overflow)
 {
 
 }
 
-FilterFir::FilterFir(const std::vector<double> &a, int max_buffer_size, bool throw_overflow):
+FilterFir::FilterFir(const std::vector<double> &a, size_t max_buffer_size, bool throw_overflow):
     SignalProcessorBuffered(max_buffer_size, throw_overflow),
     m_a(a)
 {
